add xtd::net::sockets::multicast_option class

+++ src/xtd.core/include/xtd/net/sockets/ip_v6_multicast_option.h
@@ -28,8 +28,8 @@ namespace xtd {
         ip_v6_multicast_option(const xtd::net::ip_address& group);
         /// @brief Initializes a new instance of the xtd::net::sockets::ip_v6_multicast_option class with the specified IP multicast group and the local interface address.
         /// @param group The xtd::net::ip_address of the multicast group.
-        /// @param interface_indexthe address of the interface.
-        ip_v6_multicast_option(const xtd::net::ip_address& group, uint16_t interface_index);
+        /// @param interface_index The address of the interface.
+        ip_v6_multicast_option(const xtd::net::ip_address& group, uint32_t interface_index);
 
         /// @cond
         ip_v6_multicast_option(ip_v6_multicast_option&& ip_v6_multicast_option) = default;
@@ -43,20 +43,22 @@ namespace xtd {
         const xtd::net::ip_address& group() const noexcept;
         /// @brief Sets the IP address of a multicast group.
         /// @param value An xtd::net::ip_address that contains the Internet address of a multicast group.
+        /// @return the current instance.
         ip_v6_multicast_option& group(const xtd::net::ip_address& value) noexcept;
 
         /// @brief Gets the interface index that is associated with a multicast group.
-        /// @return A uint16_t value that specifies the address of the interface.
+        /// @return A uint32_t value that specifies the address of the interface.
         /// @remarks This property specifies the interface on which data is received or sent.
-        uint16_t interface_index() const noexcept;
+        uint32_t interface_index() const noexcept;
         /// @brief Sets the interface index that is associated with a multicast group.
-        /// @param value A uint16_t value that specifies the address of the interface.
-        /// @remarks This property specifies the interface on which data is received or sent.
-        ip_v6_multicast_option& interface_index(uint16_t value) noexcept;
+        /// @param value A uint32_t value that specifies the address of the interface.
+        /// @return the current instance.
+       /// @remarks This property specifies the interface on which data is received or sent.
+        ip_v6_multicast_option& interface_index(uint32_t value) noexcept;
 
       private:
         xtd::net::ip_address group_ = xtd::net::ip_address::ip_v6_none;
-        uint16_t interface_index_ = 0;
+        uint32_t interface_index_ = 0;
       };
     }
   }
