+++ src/xtd.core/src/xtd/net/sockets/ip_v6_multicast_option.cpp
@@ -8,7 +8,7 @@ using namespace xtd::net::sockets;
 ip_v6_multicast_option::ip_v6_multicast_option(const xtd::net::ip_address& group) : group_(group) {
 }
 
-ip_v6_multicast_option::ip_v6_multicast_option(const xtd::net::ip_address& group, uint16_t interface_index) : group_(group), interface_index_(interface_index) {
+ip_v6_multicast_option::ip_v6_multicast_option(const xtd::net::ip_address& group, uint32_t interface_index) : group_(group), interface_index_(interface_index) {
 }
 
 const xtd::net::ip_address& ip_v6_multicast_option::group() const noexcept {
@@ -20,11 +20,11 @@ ip_v6_multicast_option& ip_v6_multicast_option::group(const xtd::net::ip_address
   return *this;
 }
 
-uint16_t ip_v6_multicast_option::interface_index() const noexcept {
+uint32_t ip_v6_multicast_option::interface_index() const noexcept {
   return interface_index_;
 }
 
-ip_v6_multicast_option& ip_v6_multicast_option::interface_index(uint16_t value) noexcept {
+ip_v6_multicast_option& ip_v6_multicast_option::interface_index(uint32_t value) noexcept {
   interface_index_ = value;
   return *this;
 }
